utr: Fetch and run the internal recipe for internal builders
There's now an internal recipe for the utr that should be used for
running internal builders. So this updates the utr script to
download the internal bundle and run the new recipe for chrome builders.

Since installing two different bundles in the same dir can have
conflicts, this moves the destination of the chromium bundle from
//tools/utr/.bundle/ to //tools/utr/.bundles/chromium/ and puts the
chrome bundle at //tools/utr/.bundles/chrome/.

Bug: 41492688
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5510924
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298389}

+++ tools/utr/builders_test.py
@@ -33,7 +33,8 @@ class BuilderPropsTests(unittest.TestCase):
 
   def tearDown(self):
     shutil.rmtree(self.tmp_dir)
-    shutil.rmtree(self.tmp_internal_dir)
+    if self.tmp_internal_dir.exists():
+      shutil.rmtree(self.tmp_internal_dir)
 
   def testNoProps(self):
     # Empty base dir
@@ -47,6 +48,11 @@ class BuilderPropsTests(unittest.TestCase):
 
     # Empty builder dir
     os.makedirs(self.tmp_dir.joinpath('some-bucket', 'some-builder'))
+    props, _ = builders.find_builder_props('some-bucket', 'some-builder')
+    self.assertIsNone(props)
+
+    # No src-internal checkout.
+    shutil.rmtree(self.tmp_internal_dir)
     props, _ = builders.find_builder_props('some-bucket', 'some-builder')
     self.assertIsNone(props)
